backend/app/Scraper/requirements.txt
python-dotenv
requests
beautifulsoup4
mysql-connector-python